pyproject.toml
[project]
name = "ecliptica"
version = "0.1.0"
description = "Compute and animate local solar eclipse circumstances."
readme = "README.md"
authors = [
    { name = "Michel Klueger", email = "klueger@ecoinvent.org" }
]
requires-python = ">=3.11"
dependencies = [
    "astronomy-engine>=2.1.19",
    "msgspec>=0.19.0",
    "orjson>=3.10.0",
    "typer>=0.12.5",
]

[project.optional-dependencies]
viz = [
    "manim>=0.19.0",
]

[project.scripts]
ecliptica = "ecliptica.cli:main"

[build-system]
requires = ["uv_build>=0.8.3,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=9.0.2",
    "ruff>=0.15.1",
    "ty>=0.0.17",
]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["ALL"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
